pyproject.toml
[project]
name = "tiktok-fade-in"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "moviepy>=2.2.1",
    "pillow>=11.3.0",
    "requests>=2.32.5",
]
